Clean up virtio_net
Rename vguest to virtio_net.
Remove unused variabled, remove commented code, etc.

+++ src/lib/checksum.h
@@ -21,4 +21,5 @@ void checksum_update_incremental_32(uint16_t* checksum_cell,
                                     uint32_t new_value);
 
 uint32_t pseudo_header_initial(const int8_t *buf, size_t len);
-int *prepare_packet(int8_t *buf, size_t len);
+
+uint8_t *prepare_packet(int8_t *buf, size_t len);
